% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/merge_hrv.R
\name{merge_hrv}
\alias{merge_hrv}
\title{Calculate stress}
\usage{
merge_hrv(data, hrv_data)
}
\arguments{
\item{data}{Data frame containing ECG, EEG, GPS and heart rate data.}

\item{hrv_data}{HRVData structure (RHRV package) containing data of HRV analysis.}
}
\value{
Output of the function is the modified data frame "data" with new columns HRV, LF/HF and stress added.
}
\description{
Function merges data (containing ECG, EEG and GPS values) and hrv_data (containing results of the HRV analysis), which are the input parameters.
Adds columns LF/HF and stress to data.
}
\examples{
load(system.file("extdata", "ecg.Rda", package="stressHR"))
hrv_data <- hrv_analyze(system.file("extdata", "Rsec_data.txt", package="stressHR"))
ecg <- merge_hrv(ecg, hrv_data)
}
\keyword{ECG,}
\keyword{HRV}
\keyword{calculation,}
\keyword{stress}

